requirements.txt
discord.py
discord_webhook
paho-mqtt
pytz
